separate validate_scene from read_scene
