+++ Makefile
@@ -6,7 +6,7 @@
 #    By: gsmereka <gsmereka@student.42.fr>          +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2023/01/20 18:26:17 by gde-mora          #+#    #+#              #
-#    Updated: 2023/06/30 15:19:36 by gsmereka         ###   ########.fr        #
+#    Updated: 2023/06/30 18:52:02 by gsmereka         ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -18,6 +18,7 @@ SRC 	=	src/main.c \
 			src/exit.c \
 			src/render.c \
 			src/init_data.c \
+			src/parsing/validate_scene_file.c \
 			src/parsing/read_scene_file.c \
 			src/parsing/add_config.c \
 			src/parsing/add_shape.c \
