+++ headers/miniRT.h
@@ -6,7 +6,7 @@
 /*   By: gsmereka <gsmereka@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/06/28 20:17:31 by gde-mora          #+#    #+#             */
-/*   Updated: 2023/06/30 15:21:03 by gsmereka         ###   ########.fr       */
+/*   Updated: 2023/06/30 18:53:19 by gsmereka         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -31,7 +31,8 @@
 }	t_mlx_render;*/
 
 void	render(t_data *data);
-int		read_scene_file(int argc, char *argv[], t_data *data);
+int		validate_scene_file(int argc, char *argv[], t_data *data);
+int		read_scene_file(char *file, t_data *data);
 int		exit_error(char *msg, int status, t_data *data);
 int		exit_successful(t_data *data);
 int		init_data(t_data *data);
