+++ src/main.c
@@ -6,7 +6,7 @@
 /*   By: gsmereka <gsmereka@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/06/28 20:16:42 by gde-mora          #+#    #+#             */
-/*   Updated: 2023/06/30 15:20:47 by gsmereka         ###   ########.fr       */
+/*   Updated: 2023/06/30 18:52:35 by gsmereka         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,7 +18,9 @@ int	main(int argc, char **argv)
 
 	ft_bzero(&data, sizeof(data));
 	init_data(&data);
-	read_scene_file(argc, argv, &data);
+	validate_scene_file(argc, argv, &data);
+	read_scene_file(argv[1], &data);
+	get_scene_info(&data);
 	render(&data);
 	return (0);
 }
