+++ src/parsing/check_configs.c
@@ -6,7 +6,7 @@
 /*   By: gsmereka <gsmereka@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/06/30 11:57:27 by gsmereka          #+#    #+#             */
-/*   Updated: 2023/06/30 15:19:59 by gsmereka         ###   ########.fr       */
+/*   Updated: 2023/06/30 18:48:17 by gsmereka         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,6 @@
 
 int	is_ambient_lighting(char **args)
 {
-	ft_printf("checando\n");
 	if (args && ft_strcmp(args[0], "A") == 0)
 	{
 		ft_printf("positive\n");
@@ -25,7 +24,6 @@ int	is_ambient_lighting(char **args)
 
 int	is_light(char **args)
 {
-	ft_printf("checando\n");
 	if (args && ft_strcmp(args[0], "L") == 0)
 	{
 		ft_printf("positive\n");
@@ -36,7 +34,6 @@ int	is_light(char **args)
 
 int	is_camera(char **args)
 {
-	ft_printf("checando\n");
 	if (args && ft_strcmp(args[0], "C") == 0)
 	{
 		ft_printf("positive\n");
@@ -47,7 +44,6 @@ int	is_camera(char **args)
 
 int	is_resolution(char **args)
 {
-	ft_printf("checando\n");
 	if (args && ft_strcmp(args[0], "R") == 0)
 	{
 		ft_printf("positive\n");
