+++ src/parsing/get_scene_info.c
@@ -6,7 +6,7 @@
 /*   By: gsmereka <gsmereka@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/06/30 15:13:32 by gsmereka          #+#    #+#             */
-/*   Updated: 2023/06/30 15:14:38 by gsmereka         ###   ########.fr       */
+/*   Updated: 2023/06/30 15:27:20 by gsmereka         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
