+++ src/parsing/read_scene_file.c
@@ -6,47 +6,20 @@
 /*   By: gsmereka <gsmereka@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/06/28 15:36:42 by gsmereka          #+#    #+#             */
-/*   Updated: 2023/06/30 15:16:46 by gsmereka         ###   ########.fr       */
+/*   Updated: 2023/06/30 18:52:16 by gsmereka         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../headers/miniRT.h"
 
-int	check_file_extension(char *file, t_data *data);
-int	read_file(t_data *data);
-
-int	read_scene_file(int argc, char *argv[], t_data *data)
-{
-	if (argc != 2)
-		exit_error(COUNT_ERROR, 2, data);
-	check_file_extension(argv[1], data);
-	data->scene_fd = open(argv[1], O_RDWR);
-	if (data->scene_fd < 0)
-		exit_error(OPEN_FILE_ERROR, 2, data);
-	read_file(data);
-	ft_printf("Validando\n");
-	return (0);
-}
-
-int	check_file_extension(char *file, t_data *data)
-{
-	char	*last_three_chars;
-	int		last_three_chars_index;
-
-	last_three_chars_index = ft_strlen(file) - 3;
-	if (last_three_chars_index < 0)
-		exit_error(FILE_EXTENSION_ERROR, 2, data);
-	last_three_chars = file + last_three_chars_index;
-	if (ft_strncmp(last_three_chars, ".rt", 3) != 0)
-		exit_error(FILE_EXTENSION_ERROR, 2, data);
-	return (0);
-}
-
-int	read_file(t_data *data)
+int	read_scene_file(char *file, t_data *data)
 {
 	char	*line;
 	char	**args;
 
+	data->scene_fd = open(file, O_RDWR);
+	if (data->scene_fd < 0)
+		exit_error(OPEN_FILE_ERROR, 2, data);
 	line = get_next_line(data->scene_fd);
 	while (line)
 	{
@@ -56,6 +29,6 @@ int	read_file(t_data *data)
 			add_token(&data->tokens, args);
 		line = get_next_line(data->scene_fd);
 	}
-	token_clear(&data->tokens);
+	close(data->scene_fd);
 	return (0);
 }
